source_Inductive/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.0
dill==0.3.8
filelock==3.16.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gpytorch==1.13
huggingface-hub==0.25.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jaxtyping==0.2.19
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
linear-operator==0.5.3
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyarrow==17.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
xxhash==3.5.0
yarl==1.11.1
zipp==3.20.2
